requirements.txt
streamlit
mistralai>=0.0.11
opencv-python-headless
pillow
numpy
python-dotenv